% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_data.R
\name{event_data}
\alias{event_data}
\title{Dump all records for an event data set}
\usage{
event_data(name = NULL, location = NULL, date_begin = NULL,
  date_end = NULL, filter = NULL)
}
\arguments{
\item{name}{The name of the dataset to extract}

\item{location}{(`type = "event"` and `type = "shape"` only) filter data sets with a valid geojson polygon}

\item{date_begin}{(`type = "event"` only) filter data sets for those that begin on or after this date specified in `YYYY-MM-DD` format}

\item{date_end}{(`type = "event"` only) filter data sets for those that end on or before this date specified in `YYYY-MM-DD` format}

\item{filter}{Attribute filtering, see the [documentation](http://docs.plenar.io/#attribute-filtering)}
}
\value{
A data.table
}
\description{
Dump all records for an event data set
}
\examples{
x <- event_data(name = "crimes_2001_to_present",
  filter = '{"op": "eq", "col": "primary_type", "val": "HOMICIDE"}',
  date_begin = "2016-12-01", date_end = "2016-12-31")
}
