% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{Check a GET verb call for errors}
\usage{
get_response(url, content_only = TRUE)
}
\arguments{
\item{url}{A query to the plenar.io API in the form of a character string}

\item{content_only}{A}
}
\value{
httr response object
}
\description{
Check a GET verb call for errors
}
\seealso{
\code{\link{validate_url}}
}
