% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plenar_api}
\alias{plenar_api}
\title{The URL for the Plenar.io API}
\format{A character string}
\source{
\url{http://docs.plenar.io/#welcome}
}
\usage{
plenar_api
}
\description{
The URL for the Plenar.io API
}
\keyword{datasets}
