% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_url.R
\name{validate_url}
\alias{validate_url}
\title{Basic quality checks on URL(s)}
\usage{
validate_url(urls)
}
\arguments{
\item{urls}{A character vector of URL(s)}
}
\value{
The original URL(s)
}
\description{
Basic quality checks on URL(s)
}
